fix(psynch): stop truncating the 64-bit cvsignal arguments
XNU declares syscall 304 as

  psynch_cvsignal(user_addr_t cv, uint64_t cvlsgen, uint32_t cvugen,
                  int thread_port, user_addr_t mutex, uint64_t mugen,
                  uint64_t tid, uint32_t flags)

but sys_psynch_cvsignal declared cvlsgen and mugen as uint32_t. On
arm64 the compiler therefore read w1 and w5 instead of x1 and x5 and
dropped the upper half of both; the shipped
libsystem_kernel.dylib spills the first with `stur w1, [x29, #-0xc]`.

cvlsgen carries (S << 32) | L, so every pthread_cond_signal reached
darlingserver with csgen == 0. That value is what the kext uses to
resynchronise its own copy of the condvar's S word: when a previous
drain had zeroed the wait queue, UPDATE_CVKWQ takes the
KSYN_KWF_ZEROEDOUT branch and assigns kw_sword = csgen, so the
kernel's S becomes 0 while user space's is, say, 0x401.
diff_genseq(kw_lword, kw_sword) then reports a backlog of waiters
that does not exist and ksyn_cvsignal manufactures preposts for them.

Instrumenting darlingserver over 22 `brew config` runs, 22 of 598
psynch_cvsignal operations found the kernel's S word disagreeing with
the value user space had last passed to psynch_cvwait, the kernel's
value being 0 in 21 of them.

This restores the declared ABI. It is not claimed to fix
VibeDarling/darling#99: with it applied the rate of

  PSYNCH: psync_cvwait; invalid sequence numbers

was unchanged (14 over 28 completed runs, against 22 over 53 without),
so at least one other defect drives that error.

psynch_cvbroad passed mugen where darlingserver's RPC expects the
mutex address; the kext ignores the argument, but the wrong value
should not be on the wire. Fixed alongside.

Refs: VibeDarling/darling#99

+++ darling/src/libsystem_kernel/emulation/src/xnu_syscall/bsd/impl/psynch/psynch_cvsignal.c
@@ -6,7 +6,7 @@
 
 #include <darlingserver/rpc.h>
 
-long sys_psynch_cvsignal(void* cv, uint32_t cvlsgen, uint32_t cvugen, int thread_port, void* mutex, uint32_t mugen,
+long sys_psynch_cvsignal(void* cv, uint64_t cvlsgen, uint32_t cvugen, int thread_port, void* mutex, uint64_t mugen,
 		uint64_t tid, uint32_t flags)
 {
 	uint32_t retval;
